Added map stats

+++ HeroesMatchTracker.Core/Models/StatisticsModels/StatsOverviewMaps.cs
@@ -5,6 +5,6 @@
         public string MapName { get; set; }
         public int Wins { get; set; }
         public int Losses { get; set; }
-        public decimal Winrate { get; set; }
+        public double Winrate { get; set; }
     }
 }
